+++ env.d.ts
@@ -1,0 +1,5 @@
+declare namespace NodeJS {
+  interface ProcessEnv {
+    WEATHER_API_KEY: string;
+  }
+}
